Support for dense FSP solver.
WARNING: this completely breaks compatibility with existing code.

See fsp_dense_io_example.py for an example of usage.

This only gives an example for the simple FSP expansion method.

The support_expander will be broken as it assumes the solution p
has a compress method. This is no longer true as p is merely
a dense distribution stored as a numpy array. It might be worth
having a look at the new code in cmepy/fsp/solver.py to see how
a dense probability distribution is translated between two
different domains using cmepy.state_enum.StateEnum instances.

Note that domain_state, domain_enum and p_0 are now required
to be explicitly passed when creating a cmepy.solver